fix handling single chapter stories and minor cleanup
